requirements.txt
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
langchain==0.1.0
langchain_openai
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
streamlit==1.30.0
openai==1.7.2
pipenv==2023.11.15